+++ tests/Neo.L2.Bridge.Cli.UnitTests/Usings.cs
@@ -1,0 +1,5 @@
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
+global using Neo;
+global using Neo.L2;
+global using Neo.L2.Bridge.Cli;
+global using Neo.L2.Bridge.Cli.Commands;